Updated the card queue system to be more easily expanded in the future.
Used new card queue code to apply a new trigger style of card, which will activate whenever the player takes damage.

Make it so viewers that die in a battle cannot respawn, so that cards which interact with health can have side effects.

Added Risking it all for a Friend and Shiv Storm.