Added support for One Time and Continous triggers.
One Time triggers simply occur one time before popping off the queue, refunding if they failed.

Cleaned up the Queue code.

Made it so that the Queue has built in methods to refund and clear, as needed.

Fixed the handleOnPlayerDamageLogic function to no longer break the moment a single viewer dies.

Updated DeathByAThousandPokes to prefer the attacker when aiming the Poke, now that Poke can be targeted at anyone.