Updated configuration.
Mod now shows in the in game mods list, with a config button.
Inside of the config is a button to open the config file in an exterior text editor.

Split the Bot into its own Class and namespace; and merged the chat namespace with it.

Split up Utilities into a bunch of separate classes under discordInteraction.util for organization.